requirements.txt
streamlit>=1.20.0
pydantic>=2.0.0
pyyaml>=6.0
jinja2>=3.0
python-docx>=0.8.11
docxtpl>=0.16.0
lxml>=4.9.0
pandas>=1.5.0
openpyxl>=3.0.0
